// src/generated/models/Stf_revisioncommentsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Stf_revisioncommentsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Stf_revisioncommentsstatecode = keyof typeof Stf_revisioncommentsstatecode;
export const Stf_revisioncommentsstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Stf_revisioncommentsstatuscode = keyof typeof Stf_revisioncommentsstatuscode;
export const Stf_revisioncommentsstf_attachlevel = {
  1: 'Strategy',
  2: 'Tactic',
  3: 'POC'
} as const;
export type Stf_revisioncommentsstf_attachlevel = keyof typeof Stf_revisioncommentsstf_attachlevel;
export const Stf_revisioncommentsstf_status = {
  1: 'Open',
  2: 'Resolved'
} as const;
export type Stf_revisioncommentsstf_status = keyof typeof Stf_revisioncommentsstf_status;
export const Stf_revisioncommentsstf_type = {
  1: 'Comment',
  2: 'Change Request'
} as const;
export type Stf_revisioncommentsstf_type = keyof typeof Stf_revisioncommentsstf_type;

export interface Stf_revisioncommentsBase {
  importsequencenumber?: number;
  overriddencreatedon?: string;
  statecode: Stf_revisioncommentsstatecode;
  statuscode?: Stf_revisioncommentsstatuscode;
  "stf_AttachedPOC@odata.bind"?: string;
  "stf_AttachedTactic@odata.bind"?: string;
  stf_attachlevel: Stf_revisioncommentsstf_attachlevel;
  "stf_Author@odata.bind"?: string;
  stf_name?: string;
  "stf_ParentStrategy@odata.bind"?: string;
  "stf_ResolvedBy@odata.bind"?: string;
  stf_resolvedon?: string;
  stf_response?: string;
  stf_revisioncommentid: string;
  stf_status?: Stf_revisioncommentsstf_status;
  stf_text: string;
  stf_type: Stf_revisioncommentsstf_type;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Stf_revisioncomments extends Stf_revisioncommentsBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  ownerid: string;
  owneridname: string;
  owneridtype: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  stf_attachedpocname?: string;
  stf_attachedtacticname?: string;
  stf_attachlevelname?: string;
  stf_authorname?: string;
  stf_authoryominame?: string;
  stf_parentstrategyname?: string;
  stf_resolvedbyname?: string;
  stf_resolvedbyyominame?: string;
  stf_statusname?: string;
  stf_typename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
  stf_attachedpoc?: object;
  _stf_attachedpoc_value?: string;
  stf_attachedtactic?: object;
  _stf_attachedtactic_value?: string;
  stf_author?: object;
  _stf_author_value?: string;
  stf_parentstrategy?: object;
  _stf_parentstrategy_value?: string;
  stf_resolvedby?: object;
  _stf_resolvedby_value?: string;
}